+++ src/Client/Core/DurableTaskClientOptions.cs
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.DurableTask.Converters;
+using DurableTask.Core.Serializing;
+using Dapr.DurableTask.Converters;
 
-namespace Microsoft.DurableTask.Client;
+namespace Dapr.DurableTask.Client;
 
 /// <summary>
 /// Common options for <see cref="DurableTaskClient" />.