+++ src/Client/Core/SerializedData.cs
@@ -1,9 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.DurableTask.Converters;
+using DurableTask.Core.Serializing;
+using Dapr.DurableTask.Converters;
 
-namespace Microsoft.DurableTask.Client;
+namespace Dapr.DurableTask.Client;
 
 /// <summary>
 /// Gets a type representing serialized data.