+++ test/Client/AzureManaged.Tests/DurableTaskSchedulerClientOptionsTests.cs
@@ -5,7 +5,7 @@ using Azure.Identity;
 using FluentAssertions;
 using Xunit;
 
-namespace Microsoft.DurableTask.Shared.AzureManaged.Tests;
+namespace Dapr.DurableTask.Shared.AzureManaged.Tests;
 
 public class DurableTaskSchedulerClientOptionsTests
 {